search-service/requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.34.3
pydantic==2.12.5
python-dotenv==1.0.1
openai==1.109.1
anthropic==0.84.0
llama-index==0.14.15
llama-index-vector-stores-qdrant==0.9.1
llama-index-embeddings-openai==0.5.2
llama-index-llms-anthropic==0.10.10
qdrant-client==1.17.0
